+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/query/IHierarchyQueryDAO.java
@@ -16,16 +16,18 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core;
+package org.apache.skywalking.oap.server.core.storage.query;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+import java.util.List;
+import org.apache.skywalking.oap.server.core.hierarchy.instance.InstanceHierarchyRelationTraffic;
+import org.apache.skywalking.oap.server.core.hierarchy.service.ServiceHierarchyRelationTraffic;
+import org.apache.skywalking.oap.server.core.storage.DAO;
 
-public class CoreModuleTest {
-    @Test
-    public void testOpenServiceList() {
-        CoreModule coreModule = new CoreModule();
+public interface IHierarchyQueryDAO extends DAO {
+    List<ServiceHierarchyRelationTraffic> readAllServiceHierarchyRelations() throws Exception;
 
-        Assertions.assertEquals(44, coreModule.services().length);
-    }
+    /**
+     * Return the given instance's hierarchy.
+     */
+    List<InstanceHierarchyRelationTraffic> readInstanceHierarchyRelations(String instanceId, String layer) throws Exception;
 }